cyber-agents/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==13.0
langgraph==0.2.28
langchain-core==0.3.8
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.7.2
pymongo==4.10.1
email-validator==2.2.0
